# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NovaRoute API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/internships": {
      "get": {
        "tags": [
          "Internships"
        ],
        "summary": "Browse",
        "operationId": "browse_api_internships_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "skill",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100,
              "default": "",
              "title": "Skill"
            }
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100,
              "default": "",
              "title": "Location"
            }
          },
          {
            "name": "work_mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(|remote|hybrid|onsite)$",
              "default": "",
              "title": "Work Mode"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(newest|oldest|title)$",
              "default": "newest",
              "title": "Sort"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 12,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ListingOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Internships"
        ],
        "summary": "Create",
        "operationId": "create_api_internships_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/employer/internships": {
      "get": {
        "tags": [
          "Internships"
        ],
        "summary": "Mine",
        "operationId": "mine_api_employer_internships_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(|active|flagged|closed|pending_ai_review)$",
              "default": "",
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ListingOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/internships/{identifier}": {
      "get": {
        "tags": [
          "Internships"
        ],
        "summary": "Detail",
        "operationId": "detail_api_internships__identifier__get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Internships"
        ],
        "summary": "Edit",
        "operationId": "edit_api_internships__identifier__patch",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Internships"
        ],
        "summary": "Delete",
        "operationId": "delete_api_internships__identifier__delete",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OK"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/post-internship": {
      "post": {
        "tags": [
          "Internships"
        ],
        "summary": "Create",
        "operationId": "create_api_post_internship_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/internships/{identifier}/{action}": {
      "post": {
        "tags": [
          "Internships"
        ],
        "summary": "Action",
        "operationId": "action_api_internships__identifier___action__post",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          },
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Version"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyOut": {
        "properties": {
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "website": {
            "type": "string",
            "title": "Website"
          },
          "industry": {
            "type": "string",
            "title": "Industry"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status",
            "default": "pending"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "demo_company": {
            "type": "boolean",
            "title": "Demo Company",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "company_name",
          "website",
          "industry",
          "location",
          "description"
        ],
        "title": "CompanyOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonValue": {},
...
      "ListingEdit": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 4,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 12000,
            "minLength": 30,
            "title": "Description"
          },
          "required_skills": {
            "items": {
              "type": "string",
              "maxLength": 150,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 60,
            "title": "Required Skills"
          },
          "location": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Location"
          },
          "work_mode": {
            "type": "string",
            "enum": [
              "remote",
              "hybrid",
              "onsite"
            ],
            "title": "Work Mode"
          },
          "duration": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Duration"
          },
          "stipend_min": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stipend Min"
          },
          "stipend_max": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stipend Max"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z]{3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "content_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Content Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "description",
          "required_skills",
          "location",
          "work_mode",
          "duration",
          "content_version"
        ],
        "title": "ListingEdit"
      },
      "ListingInput": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 4,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 12000,
            "minLength": 30,
            "title": "Description"
          },
          "required_skills": {
            "items": {
              "type": "string",
              "maxLength": 150,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 60,
            "title": "Required Skills"
          },
          "location": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Location"
          },
          "work_mode": {
            "type": "string",
            "enum": [
              "remote",
              "hybrid",
              "onsite"
            ],
            "title": "Work Mode"
          },
          "duration": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Duration"
          },
          "stipend_min": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stipend Min"
          },
          "stipend_max": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stipend Max"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z]{3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "description",
          "required_skills",
          "location",
          "work_mode",
          "duration"
        ],
        "title": "ListingInput"
      },
      "ListingOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "employer_id": {
            "type": "integer",
            "title": "Employer Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "required_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Skills"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "work_mode": {
            "type": "string",
            "title": "Work Mode"
          },
          "duration": {
            "type": "string",
            "title": "Duration"
          },
          "stipend_min": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stipend Min"
          },
          "stipend_max": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stipend Max"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "admin_hidden": {
            "type": "boolean",
            "title": "Admin Hidden",
            "default": false
          },
          "content_version": {
            "type": "integer",
            "title": "Content Version"
          },
          "provenance": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Provenance"
          },
          "company": {
            "$ref": "#/components/schemas/CompanyOut"
          },
          "moderation": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moderation"
          },
          "moderation_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moderation Model"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "employer_id",
          "title",
          "description",
          "required_skills",
          "location",
          "work_mode",
          "duration",
          "stipend_min",
          "stipend_max",
          "currency",
          "deadline",
          "status",
          "content_version",
          "provenance",
          "company",
          "created_at",
          "updated_at"
        ],
        "title": "ListingOut"
      },
...
      "OK": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "OK"
      },
...
      "Page_ListingOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ListingOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[ListingOut]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "Version": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version"
        ],
        "title": "Version"
      }
    }
  }
}